Collision: Added support of collisions/traces through levels. Misc: Added support of PM movements. Network/CGame: Implemented a multi-version client-game module with snap/pm/fx/messages/prediction features. Network/CGame: Added a lot of callbacks/notifications Network/Conn: Better protocol support, better snap/packets handling better multi-version support. Network/Conn: Implemented serverTime prediction with snapshots extrapolation. Network/Master: Added support of encryption. Network/Types: Fixed some errors while serializing entities and unpacking numbers. MSG: Fixed some errors while serializing data. MSG: Fixed issues with coordinates Shader: Fixed missing shaders

+++ mohpc-lib/Formats/Skel/SkeletonBones.cpp
@@ -768,7 +768,7 @@ SkelMat4 skelBone_IKshoulder::GetDirtyTransform( const skelAnimStoreFrameList_c 
 	m_cosElbowAngle = -( ( m_lowerLength * m_lowerLength + m_upperLength * m_upperLength - sinUpperAngle * sinUpperAngle ) /
 		( m_upperLength * m_lowerLength + m_upperLength * m_lowerLength ) );
 
-	length = -sqrt( 1.0f - maxLength * maxLength );
+	length = -sqrtf( 1.0f - maxLength * maxLength );
 
 	cosA = m_cachedValue[ 0 ][ 0 ];
 	sinA = m_cachedValue[ 1 ][ 0 ];
